Move Equalizer as a module(feature)
