+++ app/build.gradle
@@ -55,6 +55,7 @@ dependencies {
     implementation project(':core:downloader:youtube')
     implementation project(':core:player')
     implementation project(':feature:equalizer')
+    implementation project(':feature:playlist-editor')
 
     implementation "androidx.media:media:$media_api_version"
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:1.8.0"
