+++ app/src/main/AndroidManifest.xml
@@ -27,7 +27,7 @@
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity
-            android:name=".screens.categories.PlaylistEditorActivity"
+            android:name="com.youtubemusic.feature.playlist_editor.PlaylistEditorActivity"
             android:screenOrientation="portrait" />
         <activity
             android:name=".screens.main.MainActivity"
