+++ app/src/main/java/com/yurii/youtubemusic/screens/saved/SavedMusicFragment.kt
@@ -6,8 +6,8 @@ import androidx.fragment.app.viewModels
 import androidx.lifecycle.*
 import com.google.android.material.tabs.TabLayoutMediator
 import com.youtubemusic.core.model.MediaItemPlaylist
-import com.yurii.youtubemusic.screens.categories.PlaylistEditorActivity
 import com.youtubemusic.feature.equalizer.EqualizerActivity
+import com.youtubemusic.feature.playlist_editor.PlaylistEditorActivity
 import com.yurii.youtubemusic.R
 import com.yurii.youtubemusic.databinding.FragmentSavedMusicBinding
 import com.yurii.youtubemusic.screens.manager.DownloadManagerActivity
@@ -48,7 +48,7 @@ class SavedMusicFragment : TabFragment<FragmentSavedMusicBinding>(
     }
 
     private fun openEqualizerActivity() {
-        startActivity(Intent(requireContext(), com.youtubemusic.feature.equalizer.EqualizerActivity::class.java))
+        startActivity(Intent(requireContext(), EqualizerActivity::class.java))
     }
 
     private fun openCategoriesEditor() {
