--- app/src/main/res/values-night/colors.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <color name="colorPrimary">#222222</color>
-    <color name="colorPrimaryDark">#121212</color>
-    <color name="colorAccent">#C5C5C5</color>
-</resources>
