+++ core/common/src/main/res/layout/dialog_rename_playlist.xml
