+++ feature/playlist-editor/src/main/res/layout/activity_playlist_editor.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".screens.categories.PlaylistEditorActivity">
+    tools:context="com.youtubemusic.feature.playlist_editor.PlaylistEditorActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
@@ -45,7 +45,7 @@
             android:layout_gravity="end|bottom"
             android:layout_margin="16dp"
             android:src="@drawable/ic_add"
-            app:backgroundTint="@color/colorAccent"
+            app:backgroundTint="?attr/colorAccent"
             app:tint="@android:color/white"
             tools:ignore="ContentDescription" />
