+++ feature/playlist-editor/src/main/res/values/strings.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="dialog_confirm_deletion_playlist_title">Delete playlist</string>
+    <string name="dialog_confirm_deletion_playlist_message">Are you sure you want to delete this playlist?</string>
+    <string name="label_edit_playlists">Playlists</string>
+</resources>
