+++ settings.gradle
@@ -23,3 +23,4 @@ include ':core:model'
 include ':core:downloader:youtube'
 include ':core:player'
 include ':feature:equalizer'
+include ':feature:playlist-editor'
